move serialize package to config
License: MIT
